+++ src/hydrology.h
@@ -32,5 +32,5 @@ NumericVector soilWaterInputs(List soil, String soilFunctions, String intercepti
                               double Cm, double LgroundPAR, double LgroundSWR, 
                               double runon = 0.0,
                               bool snowpack = true, bool modifySoil = true);
-double soilFlows(List soil, NumericVector sourceSink, int nsteps = 24,
-                 bool modifySoil = true);
+NumericVector soilFlows(List soil, NumericVector sourceSink, int nsteps = 24,
+                        bool modifySoil = true);